refactor: move segment_if, map_if, load_map_file; __init__ becomes shim
Step 5 (final) of the map_if refactor plan. The remaining inline
classes and the loader move to siblings; __init__.py becomes a thin
re-export shim of ~38 lines (was 1633 lines pre-refactor).

Files added:
- pyx12/map_if/_segment.py — segment_if
- pyx12/map_if/_root.py    — map_if (root class)
- pyx12/map_if/_loader.py  — load_map_file

The plan flagged _loader.py specifically because it's the only piece
that needs `defusedxml`, `importlib.resources`, `os.path`, and
`logging`. Isolating it keeps imports lean for callers that already
have a `map_if` instance in hand.

### loop_if cleanup
The PR-4-era runtime `from . import segment_if` inside
loop_if.__init__ has been promoted to a normal top-level
`from ._segment import segment_if`. No more cycle (the dependency
graph is now linear: _base -> _element -> _composite -> _segment ->
_loop -> _root -> _loader).

### Backward-compat
The shim re-exports nine names. The package access pattern
`pyx12.map_if.X` continues to work for every X used externally —
verified by the full pytest suite (536) including
`pyx12.map_if.is_syntax_valid` referenced by test_syntax.py (added
to the shim's re-exports for that reason).

### pyx12.pyproj updated with three new entries.

### Verification
- pytest: 536 passed
- mypy --strict: clean (82 source files, was 79)
- ruff format + ruff check --select I: clean
- Smoke import: all 9 names + is_syntax_valid resolvable
- `python -m pyx12.scripts.x12valid --help` loads end-to-end

### Final picture
| Before  | After |
|---|---|
| 1 file, 1663 lines | 8 files: __init__.py (38 ln) + 7 submodules |
| 5 classes inline | 1 class per submodule, plus shared _base |
| Monolithic element_if.is_valid (179 lines) | Orchestrator + 7 helpers |
| Duplicated is_match / is_match_qual | Shared _resolve_unique_key_field helper |
| 7 repeated data_ele lookups in element_if | One cached _data_ele field |
