+++ pyx12.pyproj
@@ -73,7 +73,10 @@
     <Compile Include="pyx12\map_if\_base.py" />
     <Compile Include="pyx12\map_if\_composite.py" />
     <Compile Include="pyx12\map_if\_element.py" />
+    <Compile Include="pyx12\map_if\_loader.py" />
     <Compile Include="pyx12\map_if\_loop.py" />
+    <Compile Include="pyx12\map_if\_root.py" />
+    <Compile Include="pyx12\map_if\_segment.py" />
     <Compile Include="pyx12\map_index.py" />
     <Compile Include="pyx12\map_override.py" />
     <Compile Include="pyx12\map_walker.py" />
