+++ pyx12/map_if/__init__.py
@@ -7,707 +7,33 @@
 #
 ######################################################################
 """
-Interface to a X12N IG Map
+Interface to a X12N IG Map.
+
+This package re-exports the classes that previously lived in a single
+map_if.py module so that `import pyx12.map_if` and
+`from pyx12.map_if import X` keep working unchanged.
 """
 
 from __future__ import annotations
 
-import logging
-import os.path
-import re
-import sys
-from collections.abc import Iterator
-from importlib.resources import files as _res_files
-from typing import IO, Any, cast
-from xml.etree.ElementTree import Element
-
-import defusedxml.ElementTree as et
-
-# Intrapackage imports
-import pyx12.segment
-
-from .. import codes, dataele, validation
-from .. import path as _path
-from ..dataele import _DataEle
-from ..errors import EngineError
-from ..path import X12Path
 from ..syntax import is_syntax_valid
 from ._base import MAXINT, _required_attr, x12_node
-
-
-############################################################
-class map_if(x12_node):
-    """
-    Map file interface
-    """
-
-    pos_map: dict[int, list[Any]]
-    cur_path: str
-    param: Any
-    ext_codes: codes.ExternalCodes
-    data_elements: dataele.DataElements
-    base_name: str
-    icvn: str | None
-
-    def __init__(self, eroot: Element, param: Any, base_path: str | None = None) -> None:
-        """
-        :param eroot: ElementTree root
-        :param param: map of parameters
-        """
-        x12_node.__init__(self)
-        self.children = None  # type: ignore[assignment]
-        self.pos_map = {}
-        self.cur_path = "/transaction"
-        self.path = "/"
-        # self.cur_iter_node = self
-        self.param = param
-        # global codes
-        self.ext_codes = codes.ExternalCodes(base_path, param.get("exclude_external_codes"))
-        self.data_elements = dataele.DataElements(base_path)
-
-        self.id = eroot.get("xid")
-
-        self.name = eroot.get("name") if eroot.get("name") else eroot.findtext("name")
-        self.base_name = "transaction"
-        for e in eroot.findall("loop"):
-            loop_node = loop_if(self, self, e)
-            try:
-                self.pos_map[loop_node.pos].append(loop_node)
-            except KeyError:
-                self.pos_map[loop_node.pos] = [loop_node]
-        for e in eroot.findall("segment"):
-            seg_node = segment_if(self, self, e)
-            try:
-                self.pos_map[seg_node.pos].append(seg_node)
-            except KeyError:
-                self.pos_map[seg_node.pos] = [seg_node]
-        self.icvn = self._get_icvn()
-
-    def _get_icvn(self) -> str | None:
-        """
-        Get the Interchange version of this map
-        Map must have a first ISA segment
-        ISA12
-        """
-        ipath = "/ISA_LOOP/ISA"
-        try:
-            node = self.getnodebypath(ipath).children[11]
-            return cast(str, node.valid_codes[0])
-        except Exception:
-            return None
-
-    def debug_print(self) -> None:
-        sys.stdout.write(self.__repr__())
-        for ord1 in sorted(self.pos_map):
-            for node in self.pos_map[ord1]:
-                node.debug_print()
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, map_if):
-            return NotImplemented
-        return self.id == other.id
-
-    def __hash__(self) -> int:
-        return (self.id or "").__hash__()
-
-    def __len__(self) -> int:
-        i = 0
-        for ord1 in sorted(self.pos_map):
-            i += len(self.pos_map[ord1])
-        return i
-
-    def get_child_count(self) -> int:
-        return self.__len__()
-
-    def get_first_node(self) -> Any:
-        pos_keys = sorted(self.pos_map)
-        if len(pos_keys) > 0:
-            return self.pos_map[pos_keys[0]][0]
-        else:
-            return None
-
-    def get_first_seg(self) -> Any:
-        first = self.get_first_node()
-        if first.is_segment():
-            return first
-        else:
-            return None
-
-    def __repr__(self) -> str:
-        """
-        :rtype: string
-        """
-        return "%s\n" % (self.id)
-
-    def _path_parent(self) -> str:
-        """
-        :rtype: string
-        """
-        return os.path.basename(os.path.dirname(self.cur_path))
-
-    def get_path(self) -> str:
-        """
-        :rtype: string
-        """
-        return self.path
-
-    def get_child_node_by_idx(self, idx: int) -> Any:
-        """
-        :param idx: zero based
-        """
-        raise EngineError("map_if.get_child_node_by_idx is not a valid call")
-
-    def getnodebypath(self, spath: str) -> Any:
-        """
-        :param spath: Path string; /1000/2000/2000A/NM102-3
-        :type spath: string
-        """
-        pathl = spath.split("/")[1:]
-        if len(pathl) == 0:
-            return None
-        # logger.debug('%s %s %s' % (self.base_name, self.id, pathl[1]))
-        for ord1 in sorted(self.pos_map):
-            for child in self.pos_map[ord1]:
-                if child.id.lower() == pathl[0].lower():
-                    if len(pathl) == 1:
-                        return child
-                    else:
-                        return child.getnodebypath("/".join(pathl[1:]))
-        raise EngineError('getnodebypath failed. Path "%s" not found' % spath)
-
-    def getnodebypath2(self, path_str: str) -> Any:
-        """
-        :param path: Path string; /1000/2000/2000A/NM102-3
-        :type path: string
-        """
-        x12path = X12Path(path_str)
-        if x12path.empty():
-            return None
-        for ord1 in sorted(self.pos_map):
-            for child in self.pos_map[ord1]:
-                if child.id.upper() == x12path.loop_list[0]:
-                    if len(x12path.loop_list) == 1:
-                        return child
-                    else:
-                        del x12path.loop_list[0]
-                        return child.getnodebypath2(x12path.format())
-        raise EngineError('getnodebypath2 failed. Path "%s" not found' % path_str)
-
-    def is_map_root(self) -> bool:
-        """
-        :rtype: boolean
-        """
-        return True
-
-    def reset_child_count(self) -> None:
-        """
-        Set cur_count of child nodes to zero
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def reset_cur_count(self) -> None:
-        """
-        Set cur_count of child nodes to zero
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def __iter__(self) -> map_if:
-        return self
-
-    def loop_segment_iterator(self) -> Iterator[Any]:
-        yield self
-        for ord1 in sorted(self.pos_map):
-            for child in self.pos_map[ord1]:
-                if child.is_loop() or child.is_segment():
-                    for c in child.loop_segment_iterator():
-                        yield c
-
-
-from ._loop import loop_if  # noqa: E402, I001
-
-
-class segment_if(x12_node):
-    """
-    Segment Interface
-    """
-
-    root: Any
-    base_name: str
-    _cur_count: int
-    syntax: list[list[Any]]
-    type: str | None
-    usage: str | None
-    pos: int
-    max_use: str | None
-    repeat: str | None
-    end_tag: str | None
-
-    def __init__(self, root: Any, parent: Any, elem: Element) -> None:
-        """
-        :param parent: parent node
-        """
-
-        x12_node.__init__(self)
-        self.root = root
-        self.parent = parent
-        self.children = []
-        self.base_name = "segment"
-        self._cur_count = 0
-        self.syntax = []
-
-        self.id = elem.get("xid")
-        self.path = self.id or ""
-        self.type = elem.get("type")
-
-        self.name = elem.get("name") if elem.get("name") else elem.findtext("name")
-        self.usage = elem.get("usage") if elem.get("usage") else elem.findtext("usage")
-        self.pos = int(_required_attr(elem, "pos"))
-        self.max_use = elem.get("max_use") if elem.get("max_use") else elem.findtext("max_use")
-        self.repeat = elem.get("repeat") if elem.get("repeat") else elem.findtext("repeat")
-
-        self.end_tag = elem.get("end_tag") if elem.get("end_tag") else elem.findtext("end_tag")
-
-        for s in elem.findall("syntax"):
-            syn_list = self._split_syntax(s.text)
-            if syn_list is not None:
-                self.syntax.append(syn_list)
-
-        children_map: dict[int, Element] = {}
-        for e in elem.findall("element"):
-            seq = int(_required_attr(e, "seq"))
-            children_map[seq] = e
-
-        for e in elem.findall("composite"):
-            seq = int(_required_attr(e, "seq"))
-            children_map[seq] = e
-
-        for seq in sorted(children_map.keys()):
-            if children_map[seq].tag == "element":
-                self.children.append(element_if(self.root, self, children_map[seq]))
-            elif children_map[seq].tag == "composite":
-                self.children.append(composite_if(self.root, self, children_map[seq]))
-
-    def debug_print(self) -> None:
-        sys.stdout.write(self.__repr__())
-        for node in self.children:
-            node.debug_print()
-
-    def __repr__(self) -> str:
-        """
-        :rtype: string
-        """
-        out = '%s "%s"' % (self.id, self.name)
-        if self.usage:
-            out += "  usage: %s" % (self.usage)
-        if self.pos:
-            out += "  pos: %i" % (self.pos)
-        if self.max_use:
-            out += "  max_use: %s" % (self.max_use)
-        out += "\n"
-        return out
-
-    def get_child_node_by_idx(self, idx: int) -> Any:
-        """
-        :param idx: zero based
-        """
-        if idx >= len(self.children):
-            return None
-        else:
-            m = [c for c in self.children if c.seq == idx + 1]
-            if len(m) == 1:
-                return m[0]
-            else:
-                raise EngineError("idx %i not found in %s" % (idx, self.id))
-
-    def get_child_node_by_ordinal(self, ord: int) -> Any:
-        """
-        Get a child element or composite by the X12 ordinal
-        :param ord: one based element/composite index.  Corresponds to the map <seq> element
-        :type ord: int
-        """
-        return self.get_child_node_by_idx(ord - 1)
-
-    def getnodebypath2(self, path_str: str) -> Any:
-        """
-        Try x12 path
-
-        :param path_str: remaining path to match
-        :type path_str: string
-        :return: matching node, or None is no match
-        """
-        x12path = X12Path(path_str)
-        if x12path.empty():
-            return None
-        if x12path.ele_idx is None:
-            return self  # matched segment only
-        ele = self.get_child_node_by_ordinal(x12path.ele_idx)
-        if x12path.subele_idx is None:
-            return ele
-        return ele.get_child_node_by_ordinal(x12path.subele_idx)
-
-    def get_max_repeat(self) -> int:
-        if self.max_use is None or self.max_use == ">1":
-            return MAXINT
-        return int(self.max_use)
-
-    def get_parent(self) -> Any:
-        """
-        :return: ref to parent class instance
-        :rtype: pyx12.x12_node
-        """
-        return self.parent
-
-    def is_first_seg_in_loop(self) -> bool:
-        """
-        :rtype: boolean
-        """
-        if self is self.get_parent().get_first_seg():
-            return True
-        else:
-            return False
-
-    def is_match(self, seg: pyx12.segment.Segment) -> bool:
-        """
-        Is data segment given a match to this segment node?
-        :param seg: data segment instance
-        :return: boolean
-        :rtype: boolean
-        """
-        if seg.get_seg_id() != self.id:
-            return False
-        key = self._resolve_unique_key_field(seg.get_seg_id(), with_qual=False)
-        if key is None:
-            return True
-        child, ele_idx, subele_idx = key
-        path = f"{ele_idx:02d}-{subele_idx}" if subele_idx else f"{ele_idx:02d}"
-        return seg.get_value(path) in child.valid_codes
-
-    def is_match_qual(
-        self,
-        seg_data: pyx12.segment.Segment,
-        seg_id: str | None,
-        qual_code: str | None,
-    ) -> tuple[bool, str | None, int | None, int | None]:
-        """
-        Is segment id and qualifier a match to this segment node and to this particular segment data?
-        :param seg_data: data segment instance
-        :type seg_data: L{segment<segment.Segment>}
-        :param seg_id: data segment ID
-        :param qual_code: an ID qualifier code
-        :return: (True if a match, qual_code, element_index, subelement_index)
-        :rtype: tuple(boolean, string, int, int)
-        """
-        if seg_id != self.id:
-            return (False, None, None, None)
-        if qual_code is None:
-            return (True, None, None, None)
-        key = self._resolve_unique_key_field(seg_id, with_qual=True)
-        if key is None:
-            return (True, None, None, None)
-        child, ele_idx, subele_idx = key
-        path = f"{ele_idx:02d}-{subele_idx}" if subele_idx else f"{ele_idx:02d}"
-        if qual_code in child.valid_codes and seg_data.get_value(path) == qual_code:
-            return (True, qual_code, ele_idx, subele_idx)
-        return (False, None, None, None)
-
-    def _resolve_unique_key_field(
-        self, seg_id: str | None, *, with_qual: bool
-    ) -> tuple[Any, int, int | None] | None:
-        """
-        Locate the child node carrying this segment's qualifier (if any).
-
-        Returns ``(validating_child, ele_idx, subele_idx_or_None)`` describing
-        where to read the qualifier value from a data segment, or ``None`` if
-        the segment has no recognizable qualifier field.
-
-        ``with_qual=False`` (used by ``is_match``) accepts the AN-typed CTX
-        composite as a valid qualifier carrier; ``with_qual=True`` (used by
-        ``is_match_qual``) only honors ID-typed qualifier fields.
-        """
-        # Element at position 01 — the common case
-        if (
-            self.children[0].is_element()
-            and self.children[0].get_data_type() == "ID"
-            and self.children[0].usage == "R"
-            and len(self.children[0].valid_codes) > 0
-        ):
-            return (self.children[0], 1, None)
-        # ENT-segment carries its qualifier at element 02 (820 special case)
-        if (
-            seg_id == "ENT"
-            and self.children[1].is_element()
-            and self.children[1].get_data_type() == "ID"
-            and len(self.children[1].valid_codes) > 0
-        ):
-            return (self.children[1], 2, None)
-        # CTX-segment can have an AN-typed composite at 01-1 (999 special case);
-        # is_match_qual ignores this branch.
-        if (
-            not with_qual
-            and seg_id == "CTX"
-            and self.children[0].is_composite()
-            and self.children[0].children[0].get_data_type() == "AN"
-            and len(self.children[0].children[0].valid_codes) > 0
-        ):
-            return (self.children[0].children[0], 1, 1)
-        # General ID-typed composite at 01-1
-        if (
-            self.children[0].is_composite()
-            and self.children[0].children[0].get_data_type() == "ID"
-            and len(self.children[0].children[0].valid_codes) > 0
-        ):
-            return (self.children[0].children[0], 1, 1)
-        # HL-segment carries its qualifier at element 03
-        if (
-            seg_id == "HL"
-            and self.children[2].is_element()
-            and len(self.children[2].valid_codes) > 0
-        ):
-            return (self.children[2], 3, None)
-        return None
-
-    def guess_unique_key_id_element(self) -> Any:
-        """
-        Some segments, like REF, DTP, and DTP are duplicated.  They are matched using the value of an ID element.
-        Which element to use varies.  This function tries to find a good candidate.
-        """
-        if (
-            self.children[0].is_element()
-            and self.children[0].get_data_type() == "ID"
-            and len(self.children[0].valid_codes) > 0
-        ):
-            return self.children[0]
-        # Special Case for 820
-        elif (
-            self.id == "ENT"
-            and self.children[1].is_element()
-            and self.children[1].get_data_type() == "ID"
-            and len(self.children[1].valid_codes) > 0
-        ):
-            return self.children[1]
-        elif (
-            self.children[0].is_composite()
-            and self.children[0].children[0].get_data_type() == "ID"
-            and len(self.children[0].children[0].valid_codes) > 0
-        ):
-            return self.children[0].children[0]
-        elif (
-            self.id == "HL"
-            and self.children[2].is_element()
-            and len(self.children[2].valid_codes) > 0
-        ):
-            return self.children[2]
-        return None
-
-    def get_unique_key_id_element(self, id_val: str) -> Any:
-        """
-        Some segments, like REF, DTP, and DTP are duplicated.  They are matched using the value of an ID element.
-        Which element to use varies.  This function tries to find a good candidate, using a key value
-        """
-
-        if (
-            self.children[0].is_element()
-            and self.children[0].get_data_type() == "ID"
-            and len(self.children[0].valid_codes) > 0
-            and id_val in self.children[0].valid_codes
-        ):
-            return self.children[0]
-        # Special Case for 820
-        elif (
-            self.id == "ENT"
-            and self.children[1].is_element()
-            and self.children[1].get_data_type() == "ID"
-            and len(self.children[1].valid_codes) > 0
-            and id_val in self.children[1].valid_codes
-        ):
-            return self.children[1]
-        elif (
-            self.children[0].is_composite()
-            and self.children[0].children[0].get_data_type() == "ID"
-            and len(self.children[0].children[0].valid_codes) > 0
-            and id_val in self.children[0].children[0].valid_codes
-        ):
-            return self.children[0].children[0]
-        elif (
-            self.id == "HL"
-            and self.children[2].is_element()
-            and len(self.children[2].valid_codes) > 0
-            and id_val in self.children[2].valid_codes
-        ):
-            return self.children[2]
-        return None
-
-    def is_segment(self) -> bool:
-        """
-        :rtype: boolean
-        """
-        return True
-
-    def is_valid(self, seg_data: pyx12.segment.Segment, errh: Any) -> bool:
-        """
-        :param seg_data: data segment instance
-        :type seg_data: L{segment<segment.Segment>}
-        :param errh: instance of error_handler
-        :rtype: boolean
-        """
-        valid = True
-        child_count = self.get_child_count()
-        if len(seg_data) > child_count:
-            err_str = 'Too many elements in segment "%s" (%s). Has %i, should have %i' % (
-                self.name,
-                seg_data.get_seg_id(),
-                len(seg_data),
-                child_count,
-            )
-            ref_des = "%02i" % (child_count + 1)
-            err_value = seg_data.get_value(ref_des)
-            errh.ele_error("3", err_str, err_value, ref_des)
-            valid = False
-
-        dtype: list[str | None] = []
-        type_list: list[str | None] = []
-        for i in range(min(len(seg_data), child_count)):
-            child_node = self.get_child_node_by_idx(i)
-            if child_node.is_composite():
-                # Validate composite
-                ref_des = "%02i" % (i + 1)
-                comp_data = seg_data.get(ref_des)
-                subele_count = child_node.get_child_count()
-                if seg_data.ele_len(ref_des) > subele_count and child_node.usage != "N":
-                    subele_node = child_node.get_child_node_by_idx(subele_count + 1)
-                    err_str = 'Too many sub-elements in composite "%s" (%s)' % (
-                        subele_node.name,
-                        subele_node.refdes,
-                    )
-                    err_value = seg_data.get_value(ref_des)
-                    errh.ele_error("3", err_str, err_value, ref_des)
-                valid &= child_node.is_valid(comp_data, errh)
-            elif child_node.is_element():
-                # Validate Element
-                if (
-                    i == 1
-                    and seg_data.get_seg_id() == "DTP"
-                    and seg_data.get_value("02") in ("RD8", "D8", "D6", "DT", "TM")
-                ):
-                    dtype = [seg_data.get_value("02")]
-                if child_node.data_ele == "1250":
-                    type_list.extend(child_node.valid_codes)
-                ele_data = seg_data.get("%02i" % (i + 1))
-                if i == 2 and seg_data.get_seg_id() == "DTP":
-                    valid &= child_node.is_valid(ele_data, errh, dtype)
-                elif child_node.data_ele == "1251" and len(type_list) > 0:
-                    valid &= child_node.is_valid(ele_data, errh, type_list)
-                else:
-                    valid &= child_node.is_valid(ele_data, errh)
-
-        for i in range(min(len(seg_data), child_count), child_count):
-            # missing required elements?
-            child_node = self.get_child_node_by_idx(i)
-            valid &= child_node.is_valid(None, errh)
-
-        for syn in self.syntax:
-            (bResult, syn_err) = is_syntax_valid(seg_data, syn)
-            if not bResult:
-                syn_type = syn[0]
-                if syn_type == "E":
-                    errh.ele_error("10", syn_err, None, syn[1])
-                else:
-                    errh.ele_error("2", syn_err, None, syn[1])
-                valid &= False
-
-        return valid
-
-    def _split_syntax(self, syntax: str | None) -> list[Any] | None:
-        """
-        Split a Syntax string into a list
-        """
-        if syntax is None or syntax[0] not in ["P", "R", "C", "L", "E"]:
-            return None
-        syn: list[Any] = [syntax[0]]
-        for i in range(len(syntax[1:]) // 2):
-            syn.append(int(syntax[i * 2 + 1 : i * 2 + 3]))
-        return syn
-
-    def get_cur_count(self) -> int:
-        """
-        :return: current count
-        :rtype: int
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def incr_cur_count(self) -> None:
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def reset_cur_count(self) -> None:
-        """
-        Set cur_count of node to zero
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def set_cur_count(self, ct: int) -> None:
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def get_counts_list(self, ct_list: list[tuple[str, int]]) -> bool:
-        """
-        Build a list of (path, ct) of the current node and parents
-        Gets the node counts to apply to another map
-        :param ct_list: List to append to
-        :type ct_list: list[(string, int)]
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def loop_segment_iterator(self) -> Iterator[Any]:
-        yield self
-
-
-from ._element import element_if  # noqa: E402, I001
-from ._composite import composite_if  # noqa: E402, I001
-
-
-def load_map_file(map_file: str, param: Any, map_path: str | None = None) -> map_if:
-    """
-    Create the map object from a file
-
-    :param map_file: filename (basename) of the map xml file to load
-    :type map_file: string
-    :param map_path: Override directory containing map xml files.  If None,
-        uses package resource folder
-    :type map_path: string
-    :rtype: pyx12.map_if
-    """
-    logger = logging.getLogger("pyx12")
-    # Reject any path component in map_file to prevent traversal out of map_path
-    if map_file != os.path.basename(map_file) or os.path.isabs(map_file):
-        raise EngineError("Invalid map file name: {}".format(map_file))
-    map_fd: IO[Any]
-    if map_path is not None:
-        logger.debug("Looking for map file '{}' in map_path '{}'".format(map_file, map_path))
-        if not os.path.isdir(map_path):
-            raise OSError(2, "Map path does not exist", map_path)
-        full_path = os.path.join(map_path, map_file)
-        if not os.path.isfile(full_path):
-            raise OSError(
-                2, "Pyx12 map file '{}' does not exist in map path".format(map_file), map_path
-            )
-        map_fd = open(full_path, encoding="utf-8")
-    else:
-        logger.debug("Looking for map file '{}' in package resources".format(map_file))
-        map_fd = _res_files("pyx12").joinpath("map", map_file).open("rb")
-    with map_fd:
-        logger.debug("Create map from %s" % (map_file))
-        parser = et.XMLParser(encoding="utf-8")
-        etree = et.parse(map_fd, parser=parser)
-        return map_if(etree.getroot(), param, map_path)
-
+from ._composite import composite_if
+from ._element import element_if
+from ._loader import load_map_file
+from ._loop import loop_if
+from ._root import map_if
+from ._segment import segment_if
 
 __all__ = [
     "MAXINT",
     "_required_attr",
-    "x12_node",
-    "map_if",
-    "loop_if",
-    "segment_if",
-    "element_if",
     "composite_if",
+    "element_if",
+    "is_syntax_valid",
     "load_map_file",
+    "loop_if",
+    "map_if",
+    "segment_if",
+    "x12_node",
 ]
